requirements/base.txt
Django>=5.0
djangorestframework>=3.15
python-decouple>=3.8
Pillow>=10.0
djangorestframework-simplejwt>=5.3
django-redis>=5.4
redis>=5.0
drf-spectacular>=0.27
pytz>=2024.1
httpx>=0.27
django-modeltranslation>=0.19
channels
channels-redis
celery==5.4.0
django-celery-beat>=2.6
flower==2.0.0
daphne>=4.0  